+++ packages/isaac-typescript-definitions/src/types/classes/Entity.d.ts
@@ -328,8 +328,8 @@ declare global {
     /**
      * The color of the gibs when the entity dies.
      *
-     * Thus field is read only, so if you want to change it, you have to replace the entire thing
-     * with a new Color object.
+     * This field is read-only, so if you want to change it, you have to replace the entire thing
+     * with a new `Color` object.
      */
     SplatColor: Readonly<Color>;